CRUD

+++ src/main/java/com/adoptme/petshop/services/PetsService.java
@@ -1,7 +1,9 @@
 package com.adoptme.petshop.services;
 
 import com.adoptme.petshop.entities.Pet;
+import com.adoptme.petshop.entities.User;
 import com.adoptme.petshop.repositories.PetsRepository;
+import com.adoptme.petshop.repositories.UsersRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -12,6 +14,7 @@ import java.util.Optional;
 public class PetsService {
 
     @Autowired private PetsRepository repository;
+    @Autowired private UsersRepository usersRepository;
 
     public Pet save(Pet pet) {
         return repository.save(pet);
@@ -25,8 +28,25 @@ public class PetsService {
         return repository.findById(id);
     }
 
-    public void destroyOne(Long id) {
-        repository.deleteById(id);
+    public Optional<Pet> destroyOne(Long id) {
+        Optional<Pet> pet = repository.findById(id);
+        if (pet.isPresent()) {
+            repository.deleteById(id);
+        }
+        return pet;
+    }
+
+    public Optional<Pet> adopt(Long petId, Long userId) {
+        Optional<Pet> pet = repository.findById(petId);
+        Optional<User> user = usersRepository.findById(userId);
+        if (pet.isEmpty() || user.isEmpty()) {
+            return pet;
+        }
+        Pet foundPet = pet.get();
+        User foundUser = user.get();
+        foundPet.setOwner(foundUser);
+        repository.save(foundPet);
+        return pet;
     }
 
 }
